feat: migrate the photo-picker to home page
